requirements.txt
dotenv~=0.9.9
python-dotenv~=1.2.1
discord.py~=2.6.4
requests